fix(deposit): more reasonable deposit creation

+++ src/main/java/example/banking/deposit/dto/DepositDto.java
@@ -19,5 +19,5 @@ public class DepositDto {
     private Integer numberOfBonusesYet;  // Used to track how many bonuses have been applied
     private Integer lengthInMonths;
     private Long accountId;
-    private Double interestRate;
+    private BigDecimal interestRate;
 }
